pyproject.toml
[project]
name = "poc-opentelemetry"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "opentelemetry-api>=1.32.1",
    "opentelemetry-exporter-otlp>=1.32.1",
    "opentelemetry-sdk>=1.32.1",
    "psutil>=7.0.0",
    "sympy>=1.13.3",
]
